+++ comps/prompt_registry/arango/DockerFile
@@ -1,0 +1,41 @@
+# Copyright (C) 2024 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+
+FROM python:3.11-slim
+
+ENV LANG=C.UTF-8
+
+RUN apt-get update -y && apt-get install -y --no-install-recommends --fix-missing \
+    build-essential \
+    libgl1-mesa-glx \
+    libjemalloc-dev
+
+RUN useradd -m -s /bin/bash user && \
+    mkdir -p /home/user && \
+    chown -R user /home/user/
+
+USER user
+
+COPY comps /home/user/comps
+COPY requirements.txt /home/user/
+
+RUN pip install --no-cache-dir --upgrade pip setuptools && \
+    pip install --no-cache-dir -r /home/user/comps/prompt_registry/arango/requirements.txt && \
+    pip install --no-cache-dir -r /home/user/requirements.txt
+
+ENV PYTHONPATH=/home/user
+
+WORKDIR /home/user/comps/prompt_registry/arango
+
+ENTRYPOINT ["python", "prompt.py"]
+    
+# # Copy the Anaconda Environment
+# COPY /opt/anaconda3/envs/ArangoDB /opt/anaconda3/envs/ArangoDB
+
+# # Set PYTHONPATH to Point to the Anaconda Environment
+# ENV PYTHONPATH=$PYTHONPATH:/opt/anaconda3/envs/ArangoDB/lib/python3.11/site-packages
+
+# # Other Dockerfile Instructions
+# COPY comps /home/user/comps
+# WORKDIR /home/user/comps/prompt_registry/arango
+# ENTRYPOINT ["python", "prompt.py"]
