+++ comps/prompt_registry/arango/config.py
