+++ comps/prompt_registry/arango/requirements.txt
@@ -1,0 +1,3 @@
+python-arango
+arango
+pydantic
